analysis: highs mids lows
