+++ .gitignore
@@ -1,4 +1,5 @@
 .*
+*.map
 
 *.mp3
 *.m4a
