+++ src/bundle-standalone.js
@@ -1,2 +1,2 @@
-var mix = require('./mix');
+var mix = require('./modules/mix');
 window.heliosAudioMixer = mix;
