+++ src/modules/debug.js
@@ -14,7 +14,6 @@ debug.level = 1;
 
 // u.log(1, arg, arg, arg) -> console.log('[Mixer] arg arg arg')
 debug.log = function(lvl) {
-  console.log(lvl, arguments);
   if(lvl <= debug.level) {
     var str = '[Mixer] '
     for (var i = 1; i < arguments.length; i++)
